refactor: entity id를 BaseEntity로 이동
