+++ src/main/java/nextstep/subway/common/BaseEntity.java
@@ -1,27 +1,39 @@
 package nextstep.subway.common;
 
+import java.time.LocalDateTime;
+
+import javax.persistence.EntityListeners;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.MappedSuperclass;
+
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
 public class BaseEntity {
-    @CreatedDate
-    private LocalDateTime createdDate;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	protected Long id;
 
-    @LastModifiedDate
-    private LocalDateTime modifiedDate;
+	@CreatedDate
+	protected LocalDateTime createdDate;
 
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
+	@LastModifiedDate
+	protected LocalDateTime modifiedDate;
 
-    public LocalDateTime getModifiedDate() {
-        return modifiedDate;
-    }
+	public Long getId() {
+		return id;
+	}
+
+	public LocalDateTime getCreatedDate() {
+		return createdDate;
+	}
+
+	public LocalDateTime getModifiedDate() {
+		return modifiedDate;
+	}
 }
