+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,11 +1,10 @@
 package nextstep.subway.line.domain;
 
+import java.util.Objects;
+
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 
 import nextstep.subway.common.BaseEntity;
 import nextstep.subway.section.domain.Section;
@@ -13,9 +12,7 @@ import nextstep.subway.section.domain.Sections;
 
 @Entity
 public class Line extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+
 	@Column(name = "name", unique = true)
 	private String name;
 	@Column(name = "color")
@@ -24,7 +21,7 @@ public class Line extends BaseEntity {
 	@Embedded
 	private Sections sections = new Sections();
 
-	public Line() {
+	protected Line() {
 	}
 
 	public Line(String name, String color) {
@@ -35,10 +32,6 @@ public class Line extends BaseEntity {
 	public void update(Line line) {
 		this.name = line.getName();
 		this.color = line.getColor();
-	}
-
-	public Long getId() {
-		return id;
 	}
 
 	public String getName() {
@@ -54,6 +47,33 @@ public class Line extends BaseEntity {
 	}
 
 	public void addSection(Section section) {
-		section.changeLine(this);
+		this.sections.addSection(section);
+	}
+
+	@Override
+	public boolean equals(Object o) {
+		if (this == o)
+			return true;
+		if (o == null || getClass() != o.getClass())
+			return false;
+		Line line = (Line)o;
+		return Objects.equals(name, line.name) && Objects.equals(color, line.color) && Objects
+			.equals(sections, line.sections);
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(name, color, sections);
+	}
+
+	@Override
+	public String toString() {
+		return "Line{" +
+			"id=" + id +
+			", createdDate=" + createdDate +
+			", modifiedDate=" + modifiedDate +
+			", name='" + name + '\'' +
+			", color='" + color + '\'' +
+			'}';
 	}
 }
