+++ src/main/java/nextstep/subway/section/domain/Section.java
@@ -1,14 +1,10 @@
 package nextstep.subway.section.domain;
 
-import java.util.Objects;
 import java.util.stream.Stream;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
@@ -18,10 +14,6 @@ import nextstep.subway.station.domain.Station;
 
 @Entity
 public class Section extends BaseEntity {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "line_id")
@@ -56,10 +48,6 @@ public class Section extends BaseEntity {
 		line.getSections().addSection(this);
 	}
 
-	public Long getId() {
-		return id;
-	}
-
 	public Line getLine() {
 		return line;
 	}
@@ -78,23 +66,6 @@ public class Section extends BaseEntity {
 
 	public Stream<Station> toStationStream() {
 		return Stream.of(this.getUpStation(), this.getDownStation());
-	}
-
-	@Override
-	public boolean equals(Object o) {
-		if (this == o)
-			return true;
-		if (o == null || getClass() != o.getClass())
-			return false;
-		Section section = (Section)o;
-		return Objects.equals(id, section.id) && Objects.equals(line, section.line) && Objects
-			.equals(upStation, section.upStation) && Objects.equals(downStation, section.downStation)
-			&& Objects.equals(distance, section.distance);
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(id, line, upStation, downStation, distance);
 	}
 
 	@Override
