+++ src/main/java/nextstep/subway/station/domain/Station.java
@@ -4,17 +4,12 @@ import java.util.Objects;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 
 import nextstep.subway.common.BaseEntity;
 
 @Entity
 public class Station extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+
 	@Column(name = "name", unique = true)
 	private String name;
 
@@ -23,10 +18,6 @@ public class Station extends BaseEntity {
 
 	public Station(String name) {
 		this.name = name;
-	}
-
-	public Long getId() {
-		return id;
 	}
 
 	public String getName() {
@@ -40,19 +31,21 @@ public class Station extends BaseEntity {
 		if (o == null || getClass() != o.getClass())
 			return false;
 		Station station = (Station)o;
-		return Objects.equals(id, station.id) && Objects.equals(name, station.name);
+		return Objects.equals(name, station.name);
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(name);
 	}
 
 	@Override
 	public String toString() {
 		return "Station{" +
 			"id=" + id +
+			", createdDate=" + createdDate +
+			", modifiedDate=" + modifiedDate +
 			", name='" + name + '\'' +
 			'}';
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(id, name);
 	}
 }
